+++ Modules/Intent.Modules.Common.CSharp/Mapping/Mappings/MapChildrenMapping.cs
@@ -20,4 +20,5 @@ public class MapChildrenMapping : CSharpMappingBase
     {
         return Children.SelectMany(x => x.GetMappingStatements()).ToList();
     }
+
 }